Fix container check
The previous check failed if the image was named spotify and there were
no containers launched. Which caused the docker inspect return info
about the image instead of the containers, this is why the .Name didn't
work.

Instead of relying on this behavior we can instead explicitly do a
`docker ps` with a filter on the `name`.

This solves: #8